Added feature to index/remove ttl expiring cells to index.

+++ plugin/src/main/java/com/stratio/cassandra/lucene/IndexWriterSkinny.java
@@ -50,34 +50,42 @@ class IndexWriterSkinny extends IndexWriter {
         optionalRow = Optional.empty();
     }
 
-    /** {@inheritDoc} */
+    /**
+     * {@inheritDoc}
+     */
     @Override
     protected void delete() {
         service.delete(key);
         optionalRow = Optional.empty();
     }
 
-    /** {@inheritDoc} */
+    /**
+     * {@inheritDoc}
+     */
     @Override
     protected void index(Row row) {
         this.optionalRow = Optional.of(row);
     }
 
-    /** {@inheritDoc} */
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public void finish() {
-        optionalRow.ifPresent(row -> {
-            if (service.needsReadBeforeWrite(key, row)) {
-                UnfilteredRowIterator iterator = service.read(key, nowInSec, opGroup);
-                if (iterator.hasNext()) {
-                    row = (Row) iterator.next();
+        if (transactionType != IndexTransaction.Type.CLEANUP) {
+            optionalRow.ifPresent(row -> {
+                if (service.needsReadBeforeWrite(key, row) || (transactionType == IndexTransaction.Type.COMPACTION)) {
+                    UnfilteredRowIterator iterator = service.read(key, nowInSec, opGroup);
+                    if (iterator.hasNext()) {
+                        row = (Row) iterator.next();
+                    }
                 }
-            }
-            if (row.hasLiveData(nowInSec)) {
-                service.upsert(key, row, nowInSec);
-            } else {
-                service.delete(key);
-            }
-        });
+                if (row.hasLiveData(nowInSec)) {
+                    service.upsert(key, row, nowInSec);
+                } else {
+                    service.delete(key);
+                }
+            });
+        }
     }
 }
